+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # CHANGELOG
 
 # [Unrealeased]
-- [Improve Kitty, iTerm render]()
+- [Improve Kitty, iTerm render](https://github.com/EruEri/oyomu/pull/9)
 
 # [0.2.2-1]
 - [Re-encrypt file with the new iv](https://github.com/EruEri/oyomu/pull/8)